Vite F2: migrate integrated mocaprender onto shared avatarRenderer + holisticPipeline
- Extract MediaPipe detection (camera/video input + Holistic + skeleton overlay +
  Kalidokit solve) into src/mocap/holisticPipeline.js, emitting the same
  {riggedPose,riggedLeftHand,riggedRightHand,riggedFace} the renderer consumes.
- mocaprender/script.js is now a ~70-line wrapper: avatarRenderer + holisticPipeline
  + IPC forward (sendBoradcast / startWebServer) + #vue0 overlay. The desktop
  render-logic triplication is retired (render.js + mocaprender share avatarRenderer;
  webserv/public remains for F5).
- holisticPipeline locateFile returns an absolute fs path derived from the page URL
  (fileURLToPath): MediaPipe reads WASM/.data via Node fs, so a relative URL ENOENT'd
  from the bundled chunk. Validated working on the integrated path (VRM + FBX).
